Release 0.9.5.0 - rc2 More reliable local miner stats Some issues fixed with timings on the local machine Print the miners version to GUI, just to avoid confusion with other forks of xmr-stak Format currencies in a proper way Reset Form1.timercounter to zero on Timer1.Start Fix miner autostart if the miner is active already, was killed, but the time to re-spawn was too short, Timer2 introduced Toolstrip added for some more status information Eoorhandling more bulletproof Determine a free tcp-port on each attempt to start the Miner More reliable Miner starting, even on rewriting settings Mindmap updated

+++ Client/SnowdenAngelsSupport/SnowdenAngelsSupport/ErrorHandling.vb
@@ -29,6 +29,12 @@ Public Class ErrorHandling
 
             End If
 
+            Try
+                Form1.sts_strip_label4.Text = "Error report created"
+            Catch exex As Exception
+
+            End Try
+
         End Set
     End Property
 
@@ -148,23 +154,38 @@ Public Class ErrorHandling
 
         Try
 
-            If Form1.allowErrorTransfer = True Then
+            If Directory.Exists(errordirectory) = True Then
 
-                Dim fileEntries As String() = Directory.GetFiles(errordirectory)
+                If Form1.allowErrorTransfer = True Then
 
-                Dim fileName As String
+                    Dim fileEntries As String() = Directory.GetFiles(errordirectory)
 
-                For Each fileName In fileEntries
+                    Dim fileName As String
 
-                    If PostFile(fileName) = True Then
+                    For Each fileName In fileEntries
 
-                        File.Delete(fileName)
+                        If PostFile(fileName) = True Then
 
-                    End If
+                            File.Delete(fileName)
 
-                Next fileName
+                        End If
 
-                Return True
+                    Next fileName
+
+                    Try
+                        Form1.sts_strip_label5.Text = "Error reports sent"
+                    Catch exex As Exception
+
+                    End Try
+
+
+                    Return True
+
+                Else
+
+                    Return False
+
+                End If
 
             Else
